+++ shared/src/commonMain/kotlin/data/credentials/IdAustriaCredentialAdapter.kt
@@ -21,50 +21,35 @@ import kotlinx.serialization.json.contentOrNull
 sealed class IdAustriaCredentialAdapter(
     private val decodePortrait: (ByteArray) -> ImageBitmap?,
 ) : CredentialAdapter() {
-    override fun getAttribute(path: NormalizedJsonPath) =
-        path.segments.firstOrNull()?.let { first ->
+    override fun getAttribute(path: NormalizedJsonPath) = path.segments.firstOrNull()?.let { first ->
+        with(IdAustriaScheme.Attributes) {
             when (first) {
                 is NormalizedJsonPathSegment.NameSegment -> when (first.memberName) {
-                    IdAustriaScheme.Attributes.BPK -> Attribute.fromValue(bpk)
-                    IdAustriaScheme.Attributes.FIRSTNAME -> Attribute.fromValue(givenName)
-                    IdAustriaScheme.Attributes.LASTNAME -> Attribute.fromValue(familyName)
-                    IdAustriaScheme.Attributes.DATE_OF_BIRTH -> Attribute.fromValue(dateOfBirth)
-                    IdAustriaScheme.Attributes.PORTRAIT -> Attribute.fromValue(portraitBitmap)
-                    IdAustriaScheme.Attributes.AGE_OVER_14 -> Attribute.fromValue(ageAtLeast14)
-                    IdAustriaScheme.Attributes.AGE_OVER_16 -> Attribute.fromValue(ageAtLeast16)
-                    IdAustriaScheme.Attributes.AGE_OVER_18 -> Attribute.fromValue(ageAtLeast18)
-                    IdAustriaScheme.Attributes.AGE_OVER_21 -> Attribute.fromValue(ageAtLeast21)
-                    IdAustriaScheme.Attributes.MAIN_ADDRESS -> when (val second =
+                    BPK -> Attribute.fromValue(bpk)
+                    FIRSTNAME -> Attribute.fromValue(givenName)
+                    LASTNAME -> Attribute.fromValue(familyName)
+                    DATE_OF_BIRTH -> Attribute.fromValue(dateOfBirth)
+                    PORTRAIT -> Attribute.fromValue(portraitBitmap)
+                    AGE_OVER_14 -> Attribute.fromValue(ageAtLeast14)
+                    AGE_OVER_16 -> Attribute.fromValue(ageAtLeast16)
+                    AGE_OVER_18 -> Attribute.fromValue(ageAtLeast18)
+                    AGE_OVER_21 -> Attribute.fromValue(ageAtLeast21)
+                    MAIN_ADDRESS -> when (val second =
                         path.segments.getOrNull(1)) {
-                        is NormalizedJsonPathSegment.NameSegment -> when (second.memberName) {
-                            IdAustriaCredentialMainAddress.GEMEINDEKENNZIFFER -> Attribute.fromValue(
-                                mainAddress?.municipalityCode
-                            )
-
-                            IdAustriaCredentialMainAddress.GEMEINDEBEZEICHNUNG -> Attribute.fromValue(
-                                mainAddress?.municipalityName
-                            )
-
-                            IdAustriaCredentialMainAddress.POSTLEITZAHL -> Attribute.fromValue(
-                                mainAddress?.postalCode
-                            )
-
-                            IdAustriaCredentialMainAddress.ORTSCHAFT -> Attribute.fromValue(
-                                mainAddress?.locality
-                            )
-
-                            IdAustriaCredentialMainAddress.STRASSE -> Attribute.fromValue(
-                                mainAddress?.street
-                            )
-
-                            IdAustriaCredentialMainAddress.HAUSNUMMER -> Attribute.fromValue(
-                                mainAddress?.houseNumber
-                            )
-
-                            IdAustriaCredentialMainAddress.STIEGE -> Attribute.fromValue(mainAddress?.stair)
-                            IdAustriaCredentialMainAddress.TUER -> Attribute.fromValue(mainAddress?.door)
-                            else -> null
-                        }
+                        is NormalizedJsonPathSegment.NameSegment ->
+                            with(IdAustriaCredentialMainAddress) {
+                                when (second.memberName) {
+                                    GEMEINDEKENNZIFFER -> Attribute.fromValue(mainAddress?.municipalityCode)
+                                    GEMEINDEBEZEICHNUNG -> Attribute.fromValue(mainAddress?.municipalityName)
+                                    POSTLEITZAHL -> Attribute.fromValue(mainAddress?.postalCode)
+                                    ORTSCHAFT -> Attribute.fromValue(mainAddress?.locality)
+                                    STRASSE -> Attribute.fromValue(mainAddress?.street)
+                                    HAUSNUMMER -> Attribute.fromValue(mainAddress?.houseNumber)
+                                    STIEGE -> Attribute.fromValue(mainAddress?.stair)
+                                    TUER -> Attribute.fromValue(mainAddress?.door)
+                                    else -> null
+                                }
+                            }
 
                         null -> Attribute.fromValue(mainAddressRaw)
 
@@ -77,6 +62,7 @@ sealed class IdAustriaCredentialAdapter(
                 else -> null
             }
         }
+    }
 
     abstract val bpk: String?
     abstract val givenName: String?