+++ shared/src/commonMain/kotlin/data/credentials/MobileDrivingLicenceCredentialAttributeCategorization.kt
@@ -15,14 +15,17 @@ object MobileDrivingLicenceCredentialAttributeCategorization : CredentialAttribu
             MobileDrivingLicenceDataElements.BIRTH_DATE,
             MobileDrivingLicenceDataElements.PORTRAIT,
             MobileDrivingLicenceDataElements.PORTRAIT_CAPTURE_DATE,
-
-            MobileDrivingLicenceDataElements.NATIONALITY, // TODO: not in figma?
+            MobileDrivingLicenceDataElements.NATIONALITY,
         ).map { NormalizedJsonPath() + it to null },
 
         PersonalDataCategory.AgeData to listOf(
+            MobileDrivingLicenceDataElements.AGE_OVER_12,
+            MobileDrivingLicenceDataElements.AGE_OVER_14,
+            MobileDrivingLicenceDataElements.AGE_OVER_16,
+            MobileDrivingLicenceDataElements.AGE_OVER_18,
+            MobileDrivingLicenceDataElements.AGE_OVER_21,
             MobileDrivingLicenceDataElements.AGE_BIRTH_YEAR,
             MobileDrivingLicenceDataElements.AGE_IN_YEARS,
-            MobileDrivingLicenceDataElements.AGE_OVER_18,
         ).map { NormalizedJsonPath() + it to null },
 
         PersonalDataCategory.BirthData to listOf(